requirements.txt
boto3
requests
Flask
pyjwt
pytest
pytest-cov
